+++ R/get_varcov.R
@@ -21,12 +21,27 @@
 #'   applies to models of class `mixor`.
 #' @param complete Logical, if `TRUE`, for `aov`, returns the full
 #'   variance-covariance matrix.
-#' @inheritParams get_predicted_ci
+#' @param vcov Variance-covariance matrix used to compute uncertainty estimates (e.g., for robust standard errors). This argument accepts a covariance matrix, a function which returns a covariance matrix, or a string which identifies the function to be used to compute the covariance matrix.
+#'  * A covariance matrix
+#'  * A function which returns a covariance matrix (e.g., `stats::vcov()`)
+#'  * A string which indicates the kind of uncertainty estimates to return.
+#'    - Heteroskedasticity-consistent: `"vcovHC"`, `"HC"`, `"HC0"`, `"HC1"`, `"HC2"`, `"HC3"`, `"HC4"`, `"HC4m"`, `"HC5"`. See `?sandwich::vcovHC`
+#'    - Cluster-robust: `"vcovCR"`, `"CR0"`, `"CR1"`, `"CR1p"`, `"CR1S"`, `"CR2"`, `"CR3"`. See `?clubSandwich::vcovCR()`
+#'    - Bootstrap: `"vcovBS"`, `"xy"`, `"residual"`, `"wild"`, `"mammen"`, `"webb"`. See `?sandwich::vcovBS`
+#'    - Other `sandwich` package functions: `"vcovHAC"`, `"vcovPC"`, `"vcovCL"`, `"vcovPL"`.
+#' @param vcov_args List of arguments to be passed to the function identified by
+#'   the `vcov` argument. This function is typically supplied by the **sandwich**
+#'   or **clubSandwich** packages. Please refer to their documentation (e.g.,
+#'   `?sandwich::vcovHAC`) to see the list of available arguments.
+
 #' @param verbose Toggle warnings.
 #' @param ... Currently not used.
 #'
 #' @note `get_varcov()` tries to return the nearest positive definite matrix
-#'   in case of a negative variance-covariance matrix.
+#'   in case of negative eigenvalues of the variance-covariance matrix. This
+#'   ensures that it is still possible, for instance, to calculate standard
+#'   errors of model parameters. A message is shown when the matrix is negative
+#'   definite and a corrected matrix is returned.
 #'
 #' @return The variance-covariance matrix, as `matrix`-object.
 #'
@@ -83,7 +98,7 @@ get_varcov.default <- function(x,
                                verbose = FALSE,
                                ...)
   }
-  .process_vcov(vc, verbose)
+  .process_vcov(vc, verbose, ...)
 }
 
 #' @export
@@ -91,6 +106,10 @@ get_varcov.maxLik <- get_varcov.default
 
 #' @export
 get_varcov.HLfit <- get_varcov.default
+
+#' @export
+get_varcov.geeglm <- get_varcov.default
+
 
 
 # mlm ---------------------------------------------
@@ -126,7 +145,7 @@ get_varcov.betareg <- function(x,
     "precision" = stats::vcov(object = x, model = "precision"),
     stats::vcov(object = x)
   )
-  .process_vcov(vc, verbose)
+  .process_vcov(vc, verbose, ...)
 }
 
 
@@ -152,7 +171,7 @@ get_varcov.DirichletRegModel <- function(x,
       vc <- stats::vcov(x)
     }
   }
-  .process_vcov(vc, verbose)
+  .process_vcov(vc, verbose, ...)
 }
 
 
@@ -170,7 +189,7 @@ get_varcov.clm2 <- function(x,
 
   vc <- stats::vcov(x)
 
-  if (.is_negativ_matrix(vc)) {
+  if (.is_negativ_matrix(vc, ...)) {
     vc <- .fix_negative_matrix(vc)
   }
 
@@ -195,6 +214,7 @@ get_varcov.clmm2 <- get_varcov.clm2
 #' @export
 get_varcov.glmx <- function(x,
                             component = c("all", "conditional", "extra"),
+                            verbose = TRUE,
                             ...) {
   .check_get_varcov_dots(x, ...)
   component <- match.arg(component)
@@ -204,12 +224,12 @@ get_varcov.glmx <- function(x,
     keep <- match(find_parameters(x)[[component]], rownames(vc))
     vc <- vc[keep, keep, drop = FALSE]
   }
-  .process_vcov(vc)
+  .process_vcov(vc, verbose, ...)
 }
 
 
 #' @export
-get_varcov.pgmm <- function(x, component = c("conditional", "all"), ...) {
+get_varcov.pgmm <- function(x, component = c("conditional", "all"), verbose = TRUE, ...) {
   .check_get_varcov_dots(x, ...)
   component <- match.arg(component)
   vc <- stats::vcov(x)
@@ -218,7 +238,7 @@ get_varcov.pgmm <- function(x, component = c("conditional", "all"), ...) {
     keep <- match(find_parameters(x)[[component]], rownames(vc))
     vc <- vc[keep, keep, drop = FALSE]
   }
-  .process_vcov(vc)
+  .process_vcov(vc, verbose, ...)
 }
 
 
@@ -236,7 +256,7 @@ get_varcov.selection <- function(x,
   }
 
   # we can't check for rank-deficiency here...
-  if (.is_negativ_matrix(vc)) {
+  if (.is_negativ_matrix(vc, ...)) {
     vc <- .fix_negative_matrix(vc)
   }
   text_remove_backticks(as.matrix(vc))
@@ -246,6 +266,7 @@ get_varcov.selection <- function(x,
 #' @export
 get_varcov.mvord <- function(x,
                              component = c("all", "conditional", "thresholds", "correlation"),
+                             verbose = TRUE,
                              ...) {
   .check_get_varcov_dots(x, ...)
   component <- match.arg(component)
@@ -259,13 +280,14 @@ get_varcov.mvord <- function(x,
     keep <- match(fp, rownames(vc))
     vc <- vc[keep, keep, drop = FALSE]
   }
-  .process_vcov(vc)
+  .process_vcov(vc, verbose, ...)
 }
 
 
 #' @export
 get_varcov.mjoint <- function(x,
                               component = c("all", "conditional", "survival"),
+                              verbose = TRUE,
                               ...) {
   .check_get_varcov_dots(x, ...)
   component <- match.arg(component)
@@ -273,12 +295,15 @@ get_varcov.mjoint <- function(x,
 
   keep <- match(find_parameters(x, flatten = TRUE, component = component), rownames(vc))
   vc <- vc[keep, keep, drop = FALSE]
-  .process_vcov(vc)
+  .process_vcov(vc, verbose, ...)
 }
 
 
 #' @export
-get_varcov.mhurdle <- function(x, component = c("all", "conditional", "zi", "zero_inflated", "infrequent_purchase", "ip", "auxiliary"), ...) {
+get_varcov.mhurdle <- function(x,
+                               component = c("all", "conditional", "zi", "zero_inflated", "infrequent_purchase", "ip", "auxiliary"),
+                               verbose = TRUE,
+                               ...) {
   .check_get_varcov_dots(x, ...)
   component <- match.arg(component)
   vc <- stats::vcov(x)
@@ -288,13 +313,13 @@ get_varcov.mhurdle <- function(x, component = c("all", "conditional", "zi", "zer
 
   keep <- match(find_parameters(x, flatten = TRUE, component = component), rownames(vc))
   vc <- vc[keep, keep, drop = FALSE]
-  .process_vcov(vc)
+  .process_vcov(vc, verbose, ...)
 }
 
 
 #' @rdname get_varcov
 #' @export
-get_varcov.truncreg <- function(x, component = c("conditional", "all"), ...) {
+get_varcov.truncreg <- function(x, component = c("conditional", "all"), verbose = TRUE, ...) {
   .check_get_varcov_dots(x, ...)
   component <- match.arg(component)
   vc <- stats::vcov(x)
@@ -302,12 +327,12 @@ get_varcov.truncreg <- function(x, component = c("conditional", "all"), ...) {
   if (component == "conditional") {
     vc <- vc[1:(nrow(vc) - 1), 1:(ncol(vc) - 1), drop = FALSE]
   }
-  .process_vcov(vc)
+  .process_vcov(vc, verbose, ...)
 }
 
 
 #' @export
-get_varcov.gamlss <- function(x, component = c("conditional", "all"), ...) {
+get_varcov.gamlss <- function(x, component = c("conditional", "all"), verbose = TRUE, ...) {
   .check_get_varcov_dots(x, ...)
   component <- match.arg(component)
   vc <- suppressWarnings(stats::vcov(x))
@@ -316,7 +341,7 @@ get_varcov.gamlss <- function(x, component = c("conditional", "all"), ...) {
     cond_pars <- length(find_parameters(x)$conditional)
     vc <- as.matrix(vc)[1:cond_pars, 1:cond_pars, drop = FALSE]
   }
-  .process_vcov(vc)
+  .process_vcov(vc, verbose, ...)
 }
 
 
@@ -363,7 +388,7 @@ get_varcov.hurdle <- function(x,
     vc <- vc[keep, keep, drop = FALSE]
   }
 
-  .process_vcov(vc)
+  .process_vcov(vc, verbose, ...)
 }
 
 #' @export
@@ -375,6 +400,7 @@ get_varcov.zerocount <- get_varcov.hurdle
 #' @export
 get_varcov.zcpglm <- function(x,
                               component = c("conditional", "zero_inflated", "zi", "all"),
+                              verbose = TRUE,
                               ...) {
   .check_get_varcov_dots(x, ...)
   component <- match.arg(component)
@@ -392,7 +418,7 @@ get_varcov.zcpglm <- function(x,
     "zero_inflated" = vc[zero, zero, drop = FALSE],
     vc[c(tweedie, zero), c(tweedie, zero), drop = FALSE]
   )
-  .process_vcov(vc)
+  .process_vcov(vc, verbose, ...)
 }
 
 
@@ -404,6 +430,7 @@ get_varcov.zcpglm <- function(x,
 #' @export
 get_varcov.glmmTMB <- function(x,
                                component = c("conditional", "zero_inflated", "zi", "dispersion", "all"),
+                               verbose = TRUE,
                                ...) {
   .check_get_varcov_dots(x, ...)
   component <- match.arg(component)
@@ -415,7 +442,7 @@ get_varcov.glmmTMB <- function(x,
     "dispersion" = stats::vcov(x)[["disp"]],
     stats::vcov(x, full = TRUE)
   )
-  .process_vcov(vc)
+  .process_vcov(vc, verbose, ...)
 }
 
 
@@ -463,7 +490,7 @@ get_varcov.MixMod <- function(x,
     }
   }
 
-  .process_vcov(vc)
+  .process_vcov(vc, verbose, ...)
 }
 
 
@@ -474,14 +501,14 @@ get_varcov.MixMod <- function(x,
 
 #' @rdname get_varcov
 #' @export
-get_varcov.brmsfit <- function(x, component = "conditional", ...) {
+get_varcov.brmsfit <- function(x, component = "conditional", verbose = TRUE, ...) {
   .check_get_varcov_dots(x, ...)
   component <- match.arg(component, choices = c("all", .all_elements()))
   params <- find_parameters(x, effects = "fixed", component = component, flatten = TRUE)
   params <- gsub("^b_", "", params)
 
   vc <- stats::vcov(x)[params, params, drop = FALSE]
-  .process_vcov(vc)
+  .process_vcov(vc, verbose, ...)
 }
 
 
@@ -492,19 +519,20 @@ get_varcov.brmsfit <- function(x, component = "conditional", ...) {
 #' @export
 get_varcov.betamfx <- function(x,
                                component = c("conditional", "precision", "all"),
+                               verbose = TRUE,
                                ...) {
   .check_get_varcov_dots(x, ...)
   component <- match.arg(component)
-  get_varcov.betareg(x$fit, component = component, ...)
+  get_varcov.betareg(x$fit, component = component, verbose = verbose, ...)
 }
 
 #' @export
 get_varcov.betaor <- get_varcov.betamfx
 
 #' @export
-get_varcov.logitmfx <- function(x, ...) {
+get_varcov.logitmfx <- function(x, verbose = TRUE, ...) {
   .check_get_varcov_dots(x, ...)
-  get_varcov(x$fit, ...)
+  get_varcov(x$fit, verbose = verbose, ...)
 }
 
 #' @export
@@ -550,50 +578,62 @@ get_varcov.mediate <- function(x, ...) {
 
 #' @rdname get_varcov
 #' @export
-get_varcov.aov <- function(x, complete = FALSE, ...) {
+get_varcov.aov <- function(x, complete = FALSE, verbose = TRUE, ...) {
   .check_get_varcov_dots(x, ...)
   vc <- suppressWarnings(stats::vcov(x, complete = complete))
-  .process_vcov(vc)
+  .process_vcov(vc, verbose, ...)
 }
 
 
 #' @export
-get_varcov.ivFixed <- function(x, ...) {
+get_varcov.ivFixed <- function(x, verbose = TRUE, ...) {
   .check_get_varcov_dots(x, ...)
-  .process_vcov(x$vcov)
+  .process_vcov(x$vcov, verbose = verbose, ...)
 }
 
 
 #' @export
-get_varcov.averaging <- function(x, ...) {
+get_varcov.averaging <- function(x, verbose = TRUE, ...) {
   .check_get_varcov_dots(x, ...)
   if (is.null(attributes(x)$modelList)) {
     warning("Can't calculate covariance matrix. Please use 'fit = TRUE' in 'model.avg()'.", call. = FALSE)
   } else {
-    get_varcov.default(x)
+    get_varcov.default(x, verbose = verbose, ...)
   }
 }
 
 
 #' @export
-get_varcov.robmixglm <- function(x, ...) {
+get_varcov.robmixglm <- function(x, verbose = TRUE, ...) {
   .check_get_varcov_dots(x, ...)
   params <- find_parameters(x, flatten = TRUE)
   np <- length(params)
   vc <- x$fit@vcov[1:np, 1:np, drop = FALSE]
 
   dimnames(vc) <- list(params, params)
-  .process_vcov(vc)
+  .process_vcov(vc, verbose, ...)
 }
 
 
 #' @export
-get_varcov.Rchoice <- function(x, ...) {
+get_varcov.bife <- function(x, verbose = TRUE, ...) {
+  .check_get_varcov_dots(x, ...)
+  params <- find_parameters(x, flatten = TRUE)
+  np <- length(params)
+  vc <- stats::vcov(x)[1:np, 1:np, drop = FALSE]
+
+  dimnames(vc) <- list(params, params)
+  .process_vcov(vc, verbose, ...)
+}
+
+
+#' @export
+get_varcov.Rchoice <- function(x, verbose = TRUE, ...) {
   .check_get_varcov_dots(x, ...)
   vc <- stats::vcov(x)
   params <- find_parameters(x, flatten = TRUE)
   dimnames(vc) <- list(params, params)
-  .process_vcov(vc)
+  .process_vcov(vc, verbose, ...)
 }
 
 
@@ -608,7 +648,7 @@ get_varcov.rq <- function(x, ...) {
   rownames(vc) <- preds
   colnames(vc) <- preds
 
-  if (.is_negativ_matrix(vc)) {
+  if (.is_negativ_matrix(vc, ...)) {
     vc <- .fix_negative_matrix(vc)
   }
 
@@ -617,12 +657,12 @@ get_varcov.rq <- function(x, ...) {
 
 
 #' @export
-get_varcov.crr <- function(x, ...) {
+get_varcov.crr <- function(x, verbose = TRUE, ...) {
   .check_get_varcov_dots(x, ...)
   vc <- as.matrix(x$var)
   params <- find_parameters(x, flatten = TRUE)
   dimnames(vc) <- list(params, params)
-  .process_vcov(vc)
+  .process_vcov(vc, verbose, ...)
 }
 
 
@@ -640,7 +680,7 @@ get_varcov.crq <- function(x, ...) {
       rownames(.x) <- preds
       colnames(.x) <- preds
 
-      if (.is_negativ_matrix(.x)) {
+      if (.is_negativ_matrix(.x, ...)) {
         .x <- .fix_negative_matrix(.x)
       }
       .x
@@ -648,7 +688,7 @@ get_varcov.crq <- function(x, ...) {
     names(vc) <- sprintf("tau (%g)", unlist(lapply(sc, function(i) i$tau)))
   } else {
     vc <- as.matrix(sc$cov)
-    if (.is_negativ_matrix(vc)) {
+    if (.is_negativ_matrix(vc, ...)) {
       vc <- .fix_negative_matrix(vc)
     }
     vc <- text_remove_backticks(as.matrix(vc))
@@ -665,34 +705,34 @@ get_varcov.rqs <- get_varcov.crq
 
 
 #' @export
-get_varcov.flexsurvreg <- function(x, ...) {
+get_varcov.flexsurvreg <- function(x, verbose = TRUE, ...) {
   .check_get_varcov_dots(x, ...)
   pars <- find_parameters(x, flatten = TRUE)
   vc <- as.matrix(stats::vcov(x))[pars, pars, drop = FALSE]
-  .process_vcov(vc)
+  .process_vcov(vc, verbose, ...)
 }
 
 
 #' @export
-get_varcov.afex_aov <- function(x, ...) {
+get_varcov.afex_aov <- function(x, verbose = TRUE, ...) {
   .check_get_varcov_dots(x, ...)
-  get_varcov(x$lm, ...)
+  get_varcov(x$lm, verbose = verbose, ...)
 }
 
 
 #' @export
-get_varcov.mixed <- function(x, ...) {
+get_varcov.mixed <- function(x, verbose = TRUE, ...) {
   .check_get_varcov_dots(x, ...)
   vc <- as.matrix(stats::vcov(x$full_model))
-  .process_vcov(vc)
+  .process_vcov(vc, verbose, ...)
 }
 
 
 #' @export
-get_varcov.cpglmm <- function(x, ...) {
+get_varcov.cpglmm <- function(x, verbose = TRUE, ...) {
   .check_get_varcov_dots(x, ...)
   vc <- as.matrix(x@vcov)
-  .process_vcov(vc)
+  .process_vcov(vc, verbose, ...)
 }
 
 #' @export
@@ -701,19 +741,19 @@ get_varcov.cpglm <- get_varcov.cpglmm
 
 
 #' @export
-get_varcov.cglm <- function(x, ...) {
+get_varcov.cglm <- function(x, verbose = TRUE, ...) {
   .check_get_varcov_dots(x, ...)
   vc <- as.matrix(x$var)
-  .process_vcov(vc)
+  .process_vcov(vc, verbose, ...)
 }
 
 
 
 #' @export
-get_varcov.mle2 <- function(x, ...) {
+get_varcov.mle2 <- function(x, verbose = TRUE, ...) {
   .check_get_varcov_dots(x, ...)
   vc <- as.matrix(x@vcov)
-  .process_vcov(vc)
+  .process_vcov(vc, verbose, ...)
 }
 
 
@@ -724,12 +764,12 @@ get_varcov.mle <- get_varcov.mle2
 
 #' @rdname get_varcov
 #' @export
-get_varcov.mixor <- function(x, effects = c("all", "fixed", "random"), ...) {
+get_varcov.mixor <- function(x, effects = c("all", "fixed", "random"), verbose = TRUE, ...) {
   .check_get_varcov_dots(x, ...)
   effects <- match.arg(effects)
   params <- find_parameters(x, effects = effects, flatten = TRUE)
   vc <- as.matrix(stats::vcov(x))[params, params, drop = FALSE]
-  .process_vcov(vc)
+  .process_vcov(vc, verbose, ...)
 }
 
 
@@ -738,10 +778,10 @@ get_varcov.glmm <- get_varcov.mixor
 
 
 #' @export
-get_varcov.gamm <- function(x, ...) {
+get_varcov.gamm <- function(x, verbose = TRUE, ...) {
   .check_get_varcov_dots(x, ...)
   vc <- stats::vcov(x$gam)
-  .process_vcov(vc)
+  .process_vcov(vc, verbose, ...)
 }
 
 
@@ -754,7 +794,7 @@ get_varcov.lqmm <- function(x, ...) {
   if (length(dim(sc$Cov)) == 3) {
     vc <- lapply(1:length(x$tau), function(i) {
       .x <- sc$Cov[, , i][1:np, 1:np]
-      if (.is_negativ_matrix(.x)) {
+      if (.is_negativ_matrix(.x, ...)) {
         .x <- .fix_negative_matrix(.x)
       }
       .x
@@ -762,7 +802,7 @@ get_varcov.lqmm <- function(x, ...) {
     names(vc) <- sprintf("tau (%g)", x$tau)
   } else {
     vc <- as.matrix(sc$Cov)[1:np, 1:np]
-    if (.is_negativ_matrix(vc)) {
+    if (.is_negativ_matrix(vc, ...)) {
       vc <- .fix_negative_matrix(vc)
     }
     vc <- text_remove_backticks(as.matrix(vc))
@@ -776,58 +816,57 @@ get_varcov.lqm <- get_varcov.lqmm
 
 
 #' @export
-get_varcov.list <- function(x, ...) {
+get_varcov.list <- function(x, verbose = TRUE, ...) {
   .check_get_varcov_dots(x, ...)
   if ("gam" %in% names(x)) {
     vc <- stats::vcov(x$gam)
-    .process_vcov(vc)
+    .process_vcov(vc, verbose, ...)
   }
 }
 
 
 #' @export
-get_varcov.BBmm <- function(x, ...) {
+get_varcov.BBmm <- function(x, verbose = TRUE, ...) {
   .check_get_varcov_dots(x, ...)
   vc <- x$fixed.vcov
-  .process_vcov(vc)
+  .process_vcov(vc, verbose, ...)
 }
 
 
 #' @export
-get_varcov.BBreg <- function(x, ...) {
+get_varcov.BBreg <- function(x, verbose = TRUE, ...) {
   .check_get_varcov_dots(x, ...)
   vc <- x$vcov
-  .process_vcov(vc)
+  .process_vcov(vc, verbose, ...)
 }
 
 
 #' @export
-get_varcov.feis <- function(x, ...) {
+get_varcov.feis <- function(x, verbose = TRUE, ...) {
   .check_get_varcov_dots(x, ...)
   vc <- x$vcov
-  .process_vcov(vc)
+  .process_vcov(vc, verbose, ...)
 }
 
 
 #' @export
-get_varcov.glimML <- function(x, ...) {
+get_varcov.glimML <- function(x, verbose = TRUE, ...) {
   .check_get_varcov_dots(x, ...)
   # installed?
   check_if_installed("aod")
 
   vc <- aod::vcov(x)
-
-  .process_vcov(vc)
+  .process_vcov(vc, verbose, ...)
 }
 
 
 #' @export
-get_varcov.vglm <- function(x, ...) {
+get_varcov.vglm <- function(x, verbose = TRUE, ...) {
   .check_get_varcov_dots(x, ...)
   # installed?
   check_if_installed("VGAM")
   vc <- VGAM::vcov(x)
-  .process_vcov(vc)
+  .process_vcov(vc, verbose, ...)
 }
 
 
@@ -836,27 +875,19 @@ get_varcov.vgam <- get_varcov.vglm
 
 
 #' @export
-get_varcov.geeglm <- function(x, ...) {
-  .check_get_varcov_dots(x, ...)
-  vc <- summary(x)$cov.unscaled
-  .process_vcov(vc)
-}
-
-
-#' @export
-get_varcov.tobit <- function(x, ...) {
+get_varcov.tobit <- function(x, verbose = TRUE, ...) {
   .check_get_varcov_dots(x, ...)
   coef_names <- find_parameters(x, flatten = TRUE)
   vc <- stats::vcov(x)[coef_names, coef_names, drop = FALSE]
-  .process_vcov(vc)
+  .process_vcov(vc, verbose, ...)
 }
 
 
 #' @export
-get_varcov.lmRob <- function(x, ...) {
+get_varcov.lmRob <- function(x, verbose = TRUE, ...) {
   .check_get_varcov_dots(x, ...)
   vc <- x$cov
-  .process_vcov(vc)
+  .process_vcov(vc, verbose, ...)
 }
 
 
@@ -866,18 +897,18 @@ get_varcov.glmRob <- get_varcov.lmRob
 
 
 #' @export
-get_varcov.coxr <- function(x, ...) {
+get_varcov.coxr <- function(x, verbose = TRUE, ...) {
   .check_get_varcov_dots(x, ...)
   vc <- x$var
-  .process_vcov(vc)
+  .process_vcov(vc, verbose, ...)
 }
 
 
 #' @export
-get_varcov.gee <- function(x, ...) {
+get_varcov.gee <- function(x, verbose = TRUE, ...) {
   .check_get_varcov_dots(x, ...)
   vc <- x$naive.variance
-  .process_vcov(vc)
+  .process_vcov(vc, verbose, ...)
 }
 
 
@@ -889,8 +920,8 @@ get_varcov.LORgee <- get_varcov.gee
 # helper-functions -----------------------------------------------------
 
 
-.process_vcov <- function(vc, verbose = TRUE) {
-  if (.is_negativ_matrix(vc)) {
+.process_vcov <- function(vc, verbose = TRUE, ...) {
+  if (.is_negativ_matrix(vc, ...)) {
     vc <- .fix_negative_matrix(vc)
   }
   # fix possible missings due to rank deficient model matrix
@@ -899,13 +930,17 @@ get_varcov.LORgee <- get_varcov.gee
 }
 
 
-.is_negativ_matrix <- function(x) {
+.is_negativ_matrix <- function(x, pd_tolerance = 1e-08, ...) {
+  # allow "NULL" to skip this test
+  if (is.null(pd_tolerance)) {
+    return(FALSE)
+  }
   if (is.matrix(x) && (nrow(x) == ncol(x))) {
     rv <- tryCatch(
       {
         eigenvalues <- eigen(x, only.values = TRUE)$values
-        eigenvalues[abs(eigenvalues) < 1e-07] <- 0
-        any(eigenvalues <= 0)
+        eigenvalues[abs(eigenvalues) < pd_tolerance] <- 0
+        any(eigenvalues < 0)
       },
       error = function(e) {
         FALSE
@@ -921,6 +956,8 @@ get_varcov.LORgee <- get_varcov.gee
 
 .fix_negative_matrix <- function(m) {
   if (requireNamespace("Matrix", quietly = TRUE)) {
+    message(format_message("The variance-covariance matrix is not positive definite. Returning the nearest positive definite matrix now.",
+                           "This ensures that eigenvalues are all positive real numbers, and thereby, for instance, it is possible to calculate standard errors for all relevant parameters."))
     as.matrix(Matrix::nearPD(m)$mat)
   } else {
     m